+++ OPHD/Common.cpp
@@ -3,7 +3,7 @@
 #include "StructureManager.h"
 #include "XmlSerializer.h"
 
-#include "MapObjects/Structures/Structure.h"
+#include "MapObjects/Structure.h"
 #include "MapObjects/Structures/Warehouse.h"
 
 #include <NAS2D/Utility.h>
