+++ OPHD/StructureManager.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "MapObjects/Structures/Structure.h"
+#include "MapObjects/Structure.h"
 #include "MapObjects/Structures/Structures.h"
 
 #include <map>
