+++ OPHD/ophd.vcxproj
@@ -269,7 +269,7 @@ IF NOT "$(VcpkgInstalledDir)" == "" (
     <ClCompile Include="MapObjects\Robots\Robot.cpp" />
     <ClCompile Include="MapObjects\Structures\Factory.cpp" />
     <ClCompile Include="MapObjects\Structures\MineFacility.cpp" />
-    <ClCompile Include="MapObjects\Structures\Structure.cpp" />
+    <ClCompile Include="MapObjects\Structure.cpp" />
     <ClCompile Include="MapObjects\MapObject.cpp" />
     <ClCompile Include="UI\CheatMenu.cpp" />
     <ClCompile Include="UI\Core\Button.cpp" />
@@ -407,7 +407,7 @@ IF NOT "$(VcpkgInstalledDir)" == "" (
     <ClInclude Include="MapObjects\Structures\SolarPanelArray.h" />
     <ClInclude Include="MapObjects\Structures\SolarPlant.h" />
     <ClInclude Include="MapObjects\Structures\StorageTanks.h" />
-    <ClInclude Include="MapObjects\Structures\Structure.h" />
+    <ClInclude Include="MapObjects\Structure.h" />
     <ClInclude Include="MapObjects\Structures\Structures.h" />
     <ClInclude Include="MapObjects\Structures\FusionReactor.h" />
     <ClInclude Include="MapObjects\Structures\SurfaceFactory.h" />
